+++ trunk/NairnFEA/src/Boundary_Conditions/Constraint.cpp
@@ -16,7 +16,7 @@ int numConstraints=0;
 
 // baseNodes is contrained node, dofNum is 1 or 2 for x or y
 // must set lambda number later
-Constraint::Constraint(long baseNode,int dofNum)
+Constraint::Constraint(int baseNode,int dofNum)
 {
 	nodes.push_back(baseNode);
 	coeff.push_back((double)1.);
@@ -27,7 +27,7 @@ Constraint::Constraint(long baseNode,int dofNum)
 #pragma mark Constraint: Methods
 
 // Enter as right side of constrain equation, but add as -factor*node
-void Constraint::AddNode(long node,double factor)
+void Constraint::AddNode(int node,double factor)
 {
 	nodes.push_back(node);
 	coeff.push_back(-factor);
